requirements.txt
Package            Version
------------------ -----------
click              8.1.3
Flask              2.2.2
gunicorn           20.1.0
importlib-metadata 4.12.0
itsdangerous       2.1.2
Jinja2             3.1.2
MarkupSafe         2.1.1
numpy              1.19.5
opencv-python      4.4.0.46
Pillow             9.2.0
pip                22.2.2
pycoral            2.0.0
setuptools         63.4.1
tflite-runtime     2.5.0.post1
Werkzeug           2.2.2
wheel              0.37.1
zipp               3.8.1
